+++ spendly/mainwindow.h
@@ -14,7 +14,7 @@ public:
 
 public slots:
     void handleAuthenticationRequest(const QString& type, const QString& email, const QString& password);
-    void handleProfileUpdate(const QString& userId, const QString& username, int monthlyIncome);
+    void handleProfileUpdate(const QString& userId, const QString& username, double monthlyIncome);
     void loadHistory(const QString& userId);
 
 private:
